+++ src/commands/mod.rs
@@ -1,5 +1,6 @@
 pub mod download_agps;
 pub mod download_tracks;
+pub mod get_settings;
 pub mod info;
 pub mod list_ports;
 pub mod update;